Uploading TMP102 Fritzing Part to products
We have a new Fritzing part for SparkFun Digital Temperature Sensor Breakout - TMP102 [SEN-13314 - https://www.sparkfun.com/products/13314 ] but it was no where to be found in our repository. The one in the Fritzing library uses an older Fritzing part for the TMP102. I was able to get the Fritzing part [ 13314_sfe_digital_temperature_sensor_breakout-TMP102.fzpz ] from the engineer.

I tested it using Fritzing and it looks correct.