Add files via upload
We never had a Fritzing part for the individual 12mm buttons. I took the 12mm black button from the SIK folder and ensured that the pins were correct based on the 10302 tactile push buttons. Since we have new colored buttons for the SIK v4.0, I modified the SVG files in InkScape and pulled the colors that were used for the buttons in the SIK manual using paint.  After updating the description, I exported the files as individual Fritzing parts. =)