Qwiic I2C JST SH Connector to Breadboard Half
This Fritzing part was generated from the "Fritzing Parto-o-matic" generic male header and modified for the Qwiic JST SH connector to breadboard. However, you can duplicate this part and wire the two together to make a cable pre-terminated with female Qwiic JST connectors on both ends for illustrative purposes. This enables the ability to extend the wire length of your choice when making a circuit diagram.

- PRT-14425
- CAB-14988
- PRT-14426
- PRT-14427
- PRT-14428
- PRT-14429
- KIT-15081